cloud_functions/oauth/constant-contact-callback/requirements.txt
functions-framework==3.*
google-cloud-secret-manager==2.*
google-cloud-pubsub==2.*
requests==2.*